+++ ui/src/views/Home/widgets.js
@@ -2,59 +2,91 @@ import OrdersWidget from './widgets/OrdersWidget';
 import ConsoWidget from './widgets/ConsoWidget';
 import BillsWidget from './widgets/BillsWidget';
 import OrdersStatusesWidget from './widgets/OrdersStatusesWidget';
+import ActivationWidget from './widgets/ActWidgets/ActivationWidget';
+import PreactivationActivationWidget from './widgets/ActWidgets/PreactivationActivationWidget';
+import CanceledWidget from './widgets/ActWidgets/CanceledWidget';
+import GenericWidget from './widgets/ActWidgets/GenericWidget';
+import MassActionTable from './widgets/MassActionTable';
+import SearchLinesByIdWidget from './widgets/SearchLinesByIdWidget';
 
 export default [
   {
-    id: 'w3',
     title: 'Mes commandes',
-    description: 'Description',
+    description: '',
     checked: true,
     large: true,
+    clickable: true,
     component: OrdersWidget,
   },
   {
-    id: 'w2',
     title: 'Dernière factures',
-    description: 'Description',
+    description: '',
     checked: true,
     large: false,
+    clickable: true,
     component: BillsWidget,
   },
   {
-    id: 'w1',
     title: 'En cours de consomation',
-    description: 'Description',
+    description: '',
     checked: true,
     large: true,
+    clickable: true,
     component: ConsoWidget,
   },
   {
-    id: 'w4',
     title: 'Statut des commandes',
-    description: 'Description',
+    description: '',
     checked: true,
     large: false,
+    clickable: true,
     component: OrdersStatusesWidget,
   },
   {
-    id: 'w5',
-    title: 'Comparer 2 lignes',
-    description: 'Description',
-    checked: false,
-    large: false,
+    title: 'Acte de gestion',
+    description: '',
+    checked: true,
+    large: true,
+    component: MassActionTable,
   },
   {
-    id: 'w6',
-    title: 'Découvrir GetDiag',
-    description: 'Description',
-    checked: false,
+    title: 'Actes: Activation',
+    description: '',
+    checked: true,
     large: false,
+    component: ActivationWidget,
+    clickable: true,
   },
   {
-    id: 'w7',
-    title: 'Historique de consommation',
-    description: 'Description',
-    checked: false,
+    title: 'Actes: Préactivation / Action',
+    description: '',
+    checked: true,
     large: false,
+    component: PreactivationActivationWidget,
+    clickable: true,
+  },
+  {
+    title: 'Actes: Résiliation',
+    description: '',
+    checked: true,
+    large: false,
+    component: CanceledWidget,
+    clickable: true,
+  },
+  {
+    title: 'Statut des actes de gestion',
+    description: '',
+    checked: true,
+    large: false,
+    clickable: true,
+    component: GenericWidget,
+  },
+  {
+    title: 'Rechercher une ligne',
+    description: 'Description',
+    checked: true,
+    large: false,
+    clickable: false,
+    component: SearchLinesByIdWidget,
   },
 ];