+++ ui/src/views/Home/widgets/ActWidgets/GenericWidget.vue
@@ -1,0 +1,184 @@
+<template>
+  <WidgetBloc :widget="widget" no-padding>
+    <Indicators v-if="indicators" :meta="indicators" :on-click="onCounterClick" no-borders small />
+  </WidgetBloc>
+</template>
+
+<script>
+import Indicators from '@/components/Indicators';
+import WidgetBloc from '@/views/Home/widgets/WidgetBloc';
+
+import { countTotalForMassAction } from '@/api/massActions';
+import { currentDateMinusMounts, formattedCurrentDate, currentDateMinusDays } from '@/utils/date';
+
+export default {
+  components: {
+    Indicators,
+    WidgetBloc,
+  },
+  props: {
+    widget: Object,
+    specificFilters: {
+      type: Array,
+      default: () => [],
+    },
+  },
+  methods: {
+    onCounterClick(indicator) {
+      this.$router.push({
+        name: 'actHistory',
+        params: {
+          queryFilters: [...indicator.filters],
+        },
+      });
+    },
+  },
+
+  mounted() {
+    this.indicators = [
+      {
+        labelKey: 'indicators.getsim.actionsInProgress',
+        color: 'text-danger',
+        clickable: true,
+        total: '-',
+        filters: [
+          ...this.specificFilters,
+          {
+            id: 'filters.actStatus',
+            values: [
+              {
+                id: 'IN_PROGRESS',
+                label: this.$t('getparc.history.col.ongoing'),
+              },
+              {
+                id: 'WAITING',
+                label: this.$t('getparc.actLines.massActionsHistory.statuses.WAITING'),
+              },
+            ],
+          },
+          {
+            id: 'filters.actDateStart',
+            label: this.$t('filters.actDateStart'),
+            startDate: formattedCurrentDate(),
+            endDate: currentDateMinusMounts(6),
+          },
+        ],
+        fetch: async (indicator, partners) => {
+          return await countTotalForMassAction(indicator.filters, partners);
+        },
+      },
+      {
+        labelKey: 'indicators.getparc.lines.inprogressTwoDays',
+        color: 'text-danger',
+        clickable: true,
+        total: '-',
+        filters: [
+          ...this.specificFilters,
+          {
+            id: 'filters.actStatus',
+            values: [
+              {
+                id: 'IN_PROGRESS',
+                label: this.$t('getparc.history.col.ongoing'),
+              },
+              {
+                id: 'WAITING',
+                label: this.$t('getparc.actLines.massActionsHistory.statuses.WAITING'),
+              },
+            ],
+          },
+          {
+            id: 'filters.actDateStart',
+            label: this.$t('filters.actDateStart'),
+            endDate: currentDateMinusDays(2),
+          },
+        ],
+        fetch: async (indicator, partners) => {
+          return await countTotalForMassAction(indicator.filters, partners);
+        },
+      },
+      {
+        labelKey: 'indicators.getsim.actionsFailed',
+        color: 'text-danger',
+        clickable: true,
+        total: '-',
+        filters: [
+          ...this.specificFilters,
+          {
+            id: 'filters.actStatus',
+            values: [
+              {
+                id: 'IN_ERROR',
+                label: this.$t('getparc.actStatuses.IN_ERROR'),
+              },
+            ],
+          },
+        ],
+        fetch: async (indicator, partners) => {
+          return await countTotalForMassAction(indicator.filters, partners);
+        },
+      },
+      {
+        labelKey: 'indicators.getparc.lines.failedTwoDays',
+        color: 'text-danger',
+        clickable: true,
+        total: '-',
+        filters: [
+          ...this.specificFilters,
+          {
+            id: 'filters.actStatus',
+            values: [
+              {
+                id: 'IN_ERROR',
+                label: this.$t('getparc.actStatuses.IN_ERROR'),
+              },
+            ],
+          },
+          {
+            id: 'filters.actDateStart',
+            label: this.$t('filters.actDateStart'),
+            endDate: currentDateMinusDays(2),
+          },
+        ],
+        fetch: async (indicator, partners) => {
+          return await countTotalForMassAction(indicator.filters, partners);
+        },
+      },
+      {
+        labelKey: 'indicators.getsim.actionsPlanned',
+        color: 'text-danger',
+        clickable: true,
+        total: '-',
+        filters: [
+          ...this.specificFilters,
+          {
+            id: 'filters.actStatus',
+            values: [
+              {
+                id: 'WAITING',
+                label: this.$t('getparc.actLines.massActionsHistory.statuses.WAITING'),
+              },
+            ],
+          },
+          {
+            id: 'filters.actDateStart',
+            label: this.$t('filters.actDateStart'),
+            startDate: formattedCurrentDate(),
+          },
+        ],
+        fetch: async (indicator, partners) => {
+          return await countTotalForMassAction(indicator.filters, partners);
+        },
+      },
+    ];
+  },
+
+  data() {
+    return {
+      indicators: undefined,
+    };
+  },
+};
+</script>
+
+<style lang="scss" scoped></style>